fix(data): address PR #1 review
- Drop dangling :class:`PairedDataset` ref in data/__init__.py — class
  no longer exists.
- Drop speculative concrete-benchmark enumeration in
  evaluation/__init__.py (subclass-enumeration anti-pattern).
- Resolve rubric contract drift: the test fake's score() now uses a
  task-specific scoring instruction string instead of self.rubric, to
  match the docstring contract that the rubric is for the
  self-rewarding labeller only.
- Tighten dtype: torch.zeros(pad_len, dtype=torch.long) so masks stay
  int64 throughout the padding path, not just after the trailing
  .long() cast.
- Reframe EvalDataset.dataset docstring: drop overclaimed "read-only"
  wording (HF Dataset is mutable) and emphasise the deterministic-row-
  ordering contract that's actually load-bearing.

+++ consistency_em/data/paired_dataset.py
@@ -103,7 +103,9 @@ class PairedDataCollator:
                         ]
                     )
                 )
-                padded_masks.append(torch.cat([torch.as_tensor(seq_mask), torch.zeros(pad_len)]))
+                padded_masks.append(
+                    torch.cat([torch.as_tensor(seq_mask), torch.zeros(pad_len, dtype=torch.long)])
+                )
             else:
                 padded_ids.append(torch.as_tensor(seq_ids))
                 padded_masks.append(torch.as_tensor(seq_mask))
